Improved/fixed code to prevent same notification to be fired more than once
It was reliant on the latest message to ensure the relevant notification was sent. If another message was added in the meantime (like a timeout message) then the notification would fire again even though the next threshold wasn't reached yet.

Now uses a list to remember the steps that were notified, which is cleared when certificate expiration date changes.